Tests : Decorate existing performance tests with PerformanceTestMethod
Also move SpeedTest methods to GraphComponentTest, since they are testing GraphComponent speed.
